client-api: Rename discover module to discovery
Align with federation-api
