+++ crates/ruma-client-api/src/discovery.rs
