+++ crates/ruma-client-api/src/discovery/discover_homeserver.rs
