+++ crates/ruma-client-api/tests/headers.rs
@@ -1,7 +1,7 @@
 #![cfg(feature = "client")]
 
 use http::HeaderMap;
-use ruma_client_api::discover::discover_homeserver;
+use ruma_client_api::discovery::discover_homeserver;
 use ruma_common::api::{MatrixVersion, OutgoingRequest as _, SendAccessToken};
 
 #[test]
