+++ crates/ruma-client/src/client/builder.rs
@@ -1,6 +1,6 @@
 use std::sync::{Arc, Mutex};
 
-use ruma_client_api::discover::get_supported_versions;
+use ruma_client_api::discovery::get_supported_versions;
 use ruma_common::api::{MatrixVersion, SendAccessToken};
 
 use super::{Client, ClientData};
